refactor mobile view
