--- services/ui-src/src/components/pages/Dashboard/DashboardProgramListMobile.tsx
@@ -1,125 +1,0 @@
-// components
-import { Box, Button, Flex, Image, Text } from "@chakra-ui/react";
-import { Spinner } from "@cmsgov/design-system";
-// utils
-import { AnyObject, ReportMetadataShape } from "types";
-import { convertDateUtcToEt } from "utils";
-// assets
-import editIcon from "assets/icons/icon_edit_square_gray.png";
-
-export const MobileDashboardList = ({
-  reportsByState,
-  reportId,
-  openAddEditReportModal,
-  enterSelectedReport,
-  archiveReport,
-  archiving,
-  sxOverride,
-  isStateLevelUser,
-  isAdmin,
-}: MobileDashboardListProps) => (
-  <>
-    {reportsByState.map((report: AnyObject) => (
-      <Box data-testid="mobile-row" sx={sx.mobileTable} key={report.id}>
-        <Box sx={sx.labelGroup}>
-          <Text sx={sx.label}>Program name</Text>
-          <Flex alignContent="flex-start">
-            {isStateLevelUser && (
-              <Box sx={sxOverride.editProgram}>
-                <button onClick={() => openAddEditReportModal(report)}>
-                  <Image
-                    src={editIcon}
-                    data-testid="mobile-edit-program"
-                    alt="Edit Program"
-                  />
-                </button>
-              </Box>
-            )}
-            <Text sx={sxOverride.reportNameText}>{report.reportName}</Text>
-          </Flex>
-        </Box>
-        <Box sx={sx.labelGroup}>
-          <Flex alignContent="flex-start">
-            <Box sx={sx.editDate}>
-              <Text sx={sx.label}>Due date</Text>
-              <Text>{convertDateUtcToEt(report.dueDate)}</Text>
-            </Box>
-            <Box>
-              <Text sx={sx.label}>Last edited</Text>
-              <Text>{convertDateUtcToEt(report.lastAltered)}</Text>
-            </Box>
-          </Flex>
-        </Box>
-        <Box sx={sx.labelGroup}>
-          <Text sx={sx.label}>Edited by</Text>
-          <Text>{report?.lastAlteredBy || "-"}</Text>
-        </Box>
-        <Box sx={sx.labelGroup}>
-          <Text sx={sx.label}>Status</Text>
-          <Text>{report?.archived ? "Archived" : report?.status}</Text>
-        </Box>
-        <Flex alignContent="flex-start" gap={2}>
-          <Box sx={sxOverride.editReportButtonCell}>
-            <Button
-              variant="outline"
-              onClick={() => enterSelectedReport(report)}
-              isDisabled={report?.archived}
-            >
-              Enter
-            </Button>
-          </Box>
-          <Box sx={sxOverride.deleteReportCell}>
-            {isAdmin && (
-              <Button
-                variant="link"
-                sx={sxOverride.archiveReportButton}
-                onClick={() => archiveReport(report)}
-              >
-                {archiving && reportId === report.id ? (
-                  <Spinner size="small" />
-                ) : report?.archived ? (
-                  "Unarchive"
-                ) : (
-                  "Archive"
-                )}
-              </Button>
-            )}
-          </Box>
-        </Flex>
-      </Box>
-    ))}
-  </>
-);
-
-interface MobileDashboardListProps {
-  reportsByState: ReportMetadataShape[];
-  reportId: string | undefined;
-  openAddEditReportModal: Function;
-  enterSelectedReport: Function;
-  archiveReport: Function;
-  archiving: boolean;
-  lockReport?: Function | undefined;
-  locking?: boolean | undefined;
-  isAdmin: boolean;
-  isStateLevelUser: boolean;
-  sxOverride: AnyObject;
-}
-
-const sx = {
-  mobileTable: {
-    padding: "1rem 0",
-    borderBottom: "1px solid",
-    borderColor: "palette.gray_light",
-  },
-  labelGroup: {
-    marginBottom: "0.5rem",
-  },
-  label: {
-    fontSize: "sm",
-    fontWeight: "bold",
-    color: "palette.gray_medium",
-  },
-  editDate: {
-    marginRight: "3rem",
-  },
-};
